Rename library to scala-tutorial
